refactor(web): retire pi-web-ui — drop deps + retire old PiChat (#1830) (#1833)
* refactor(web): repoint / route to new chat UI (#1830)

Renames PiChatV2 → PiChat at the / route, deletes the old
1091-line pi-web-ui-based PiChat, and removes the /chat-v2 route
+ "chat-v2 preview" badge. The new ai-elements-based chat is now
the default.

Part of #1808.

* refactor(web): drop pi-web-ui imports + CSS overrides (#1830)

Part of #1808 (PR 7). The new chat UI no longer renders pi-web-ui's
`<agent-interface>` / `<message-editor>` web components, so:

- delete the `@/adapters/rara-storage` adapter, `lib/synthetic-model`,
  `pages/pi-chat-messages`, and `tools/turn-chip-card` — every consumer
  of these modules went away with PiChatV2's promotion to `/`
- prune `@/adapters/rara-stream` down to `PublicWebEvent` + `buildWsUrl`,
  the only entry points the new `applyRaraEvent` reducer and
  `live-run-store` actually consume
- rip every pi-web-ui-targeted CSS rule out of `index.css`
  (`assistant-message`, `user-message-container`, `message-editor`,
  `agent-interface`, `voice-float`, `rara-live-slot`, the
  `.hidden { display: none !important }` ArtifactsPanel band-aid, etc.)
- inline the design tokens that pi-web-ui's `app.css` used to provide
  (`--background` / `--foreground` / `--card` / `--primary` / etc.) so
  Tailwind utilities + `@theme` re-export keep resolving without the
  upstream package
- rename `.rara-chat-v2` → `.rara-chat` (heading serif rules) and apply
  it to PiChat's root so markdown inside `RenderPart`'s `.prose` keeps
  the Fraunces-headed look the v2 surface had

The pi-web-ui npm package is uninstalled in a follow-up commit.

Closes #1830

* refactor(web): uninstall pi-web-ui npm deps (#1830)

Part of #1808 (PR 7). Now that no source file imports
`@mariozechner/pi-web-ui`, `@mariozechner/pi-agent-core`, or
`@mariozechner/pi-ai`, drop them from `package.json` so the bundle
stops pulling lit, mini-lit, and the upstream agent runtime.

Bundle delta (vite production build, sum of `dist/assets/*.js`):
  before: 17,929.6 KB
  after:  13,411.8 KB
  saved:    4,517.8 KB (-25.2%)

Closes #1830

* test(web): pin sidebar session locators in chat e2e (#1830)

The new chat header at `/` displays the active session's title (e.g.
"Design review" once auto-selected on mount), which collided with the
sidebar's session button text and tripped Playwright strict mode:

    strict mode violation: getByText('Design review') resolved to 2

Pin the assertion to the sidebar's button role so the heading text is
unambiguous; both tests still verify the same thing — the sidebar
renders a row per seeded session.

Closes #1830

* docs(web): refresh stale PiChatV2 / chat-v2 comment references (#1830)

Picked up in code review of #1830 — three doc comments and one AGENT.md
entry still referenced the v2 transitional names + pi-web-ui scroll
viewport. Update them so the surface code reads consistently with the
post-retirement reality.

No behaviour change.

Closes #1830

---------
