+++ web/src/hooks/AGENT.md
@@ -10,7 +10,7 @@ Real files in this directory (one line each):
 
 - `loading-hints.ts` — poetic placeholder strings + `randomLoadingHint()`, mirrors the Telegram channel's loading copy so both feel cohesive.
 - `use-dock-store.ts` — in-memory dock state machine (sessions, blocks, mutations, history) for the agent dock UI.
-- `use-live-card-height.ts` — `ResizeObserver`-driven publisher of the in-progress agent live-card height to a CSS custom property on `<main>`, so pi-web-ui's scroll viewport reserves padding for the floating overlay.
+- `use-live-card-height.ts` — `ResizeObserver`-driven publisher of the in-progress agent live-card height to a CSS custom property on `<main>`. Currently unused after pi-web-ui's retirement; kept around as a building block in case the new chat ever reintroduces a floating in-progress overlay.
 - `use-local-storage.ts` — typed `useState`-style wrapper over `window.localStorage` with JSON serialization.
 - `use-server-status.ts` — context hook exposing `{ isOnline, isChecking }` for the global server-status banner.
 - `use-session-delete.ts` — wires the pure `decidePostDeleteAction` helper into the chat page's switch/create-new side effects so the dispatch is unit-testable.
