--- web/src/pages/__tests__/pi-chat-messages.test.ts
@@ -1,395 +1,0 @@
-/*
- * Copyright 2025 Rararulab
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import type { AgentMessage } from '@mariozechner/pi-agent-core';
-import type { AssistantMessage } from '@mariozechner/pi-ai';
-import { describe, expect, it } from 'vitest';
-
-import {
-  aggregateTurnToolCalls,
-  assistantSeqByRef,
-  finalAssistantIndices,
-  isFirstAssistantOfTurn,
-  messagesForArtifactReconstruction,
-  toAgentMessages,
-  toolResultByCallId,
-} from '../pi-chat-messages';
-
-/**
- * Build the `turnHosts` set that production code derives from
- * {@link aggregateTurnToolCalls}. Tests pass an empty results map since
- * the host-picking logic depends only on the message shape, not on
- * whether results have landed yet.
- */
-function hostsOf(list: readonly AgentMessage[]): ReadonlySet<AssistantMessage> {
-  return new Set(aggregateTurnToolCalls(list, new Map()).keys());
-}
-
-import type { ChatMessageData, ChatToolCallData } from '@/api/types';
-
-const ISO = '2025-01-01T00:00:00Z';
-
-function user(seq: number, text = 'hi'): ChatMessageData {
-  return { seq, role: 'user', content: text, created_at: ISO };
-}
-
-function assistantText(seq: number, text = 'reply'): ChatMessageData {
-  return { seq, role: 'assistant', content: text, created_at: ISO };
-}
-
-function assistantToolCall(seq: number, toolName = 'do_thing'): ChatMessageData {
-  const call: ChatToolCallData = {
-    id: `tc-${seq}`,
-    name: toolName,
-    arguments: {},
-  };
-  return {
-    seq,
-    role: 'assistant',
-    content: '',
-    tool_calls: [call],
-    created_at: ISO,
-  };
-}
-
-function assistantMultiToolCall(seq: number, toolNames: string[]): ChatMessageData {
-  const calls: ChatToolCallData[] = toolNames.map((name, i) => ({
-    id: `tc-${seq}-${i}`,
-    name,
-    arguments: {},
-  }));
-  return {
-    seq,
-    role: 'assistant',
-    content: '',
-    tool_calls: calls,
-    created_at: ISO,
-  };
-}
-
-function toolResult(seq: number, callSeq: number, toolName = 'do_thing'): ChatMessageData {
-  return {
-    seq,
-    role: 'tool_result',
-    content: 'ok',
-    tool_call_id: `tc-${callSeq}`,
-    tool_name: toolName,
-    created_at: ISO,
-  };
-}
-
-function isRegistered(msg: AssistantMessage | undefined): boolean {
-  if (!msg) throw new Error('expected an assistant message');
-  return assistantSeqByRef.get(msg) !== undefined;
-}
-
-function expectAssistant(msg: AssistantMessage | undefined): AssistantMessage {
-  if (!msg) throw new Error('expected an assistant message');
-  return msg;
-}
-
-function assistants(out: ReturnType<typeof toAgentMessages>): AssistantMessage[] {
-  return out.filter((m): m is AssistantMessage => m.role === 'assistant');
-}
-
-describe('finalAssistantIndices', () => {
-  it('marks the only assistant of a single complete turn', () => {
-    const msgs = [user(1), assistantText(2)];
-    expect(finalAssistantIndices(msgs)).toEqual(new Set([1]));
-  });
-
-  it('marks only the last assistant when intermediate ones are tool-call-only', () => {
-    const msgs = [user(1), assistantToolCall(2), assistantText(3)];
-    expect(finalAssistantIndices(msgs)).toEqual(new Set([2]));
-  });
-
-  it('handles tool_result interleave between intermediate and final assistant', () => {
-    const msgs = [user(1), assistantToolCall(2), toolResult(3, 2), assistantText(4)];
-    expect(finalAssistantIndices(msgs)).toEqual(new Set([3]));
-  });
-
-  it('marks the final assistant of every turn across multiple turns', () => {
-    const msgs = [user(1), assistantText(2), user(3), assistantToolCall(4), assistantText(5)];
-    expect(finalAssistantIndices(msgs)).toEqual(new Set([1, 4]));
-  });
-
-  it('marks the trailing tool-call-only assistant on an open (incomplete) turn', () => {
-    const msgs = [user(1), assistantToolCall(2)];
-    expect(finalAssistantIndices(msgs)).toEqual(new Set([1]));
-  });
-
-  it('treats a no-op user with no assistant as inert and marks only the next turn', () => {
-    const msgs = [user(1), user(2), assistantText(3)];
-    expect(finalAssistantIndices(msgs)).toEqual(new Set([2]));
-  });
-
-  it('marks the last tool-call-only assistant on an open turn with multiple tool-call assistants', () => {
-    const msgs = [user(1), assistantToolCall(2), assistantToolCall(3)];
-    expect(finalAssistantIndices(msgs)).toEqual(new Set([2]));
-  });
-});
-
-describe('toAgentMessages — trace button registration (#1672)', () => {
-  it('registers the lone assistant of a [user, assistant(text)] turn', () => {
-    const out = toAgentMessages([user(1), assistantText(2)]);
-    const [a] = assistants(out);
-    expect(isRegistered(a)).toBe(true);
-    expect(assistantSeqByRef.get(expectAssistant(a))).toBe(2);
-  });
-
-  it('registers only the final assistant in [user, assistant(tool_call), assistant(text)]', () => {
-    const out = toAgentMessages([user(1), assistantToolCall(2), assistantText(3)]);
-    const [intermediate, final] = assistants(out);
-    expect(isRegistered(intermediate)).toBe(false);
-    expect(isRegistered(final)).toBe(true);
-    expect(assistantSeqByRef.get(expectAssistant(final))).toBe(3);
-  });
-
-  it('registers only the final assistant when a tool_result is interleaved', () => {
-    const out = toAgentMessages([
-      user(1),
-      assistantToolCall(2),
-      toolResult(3, 2),
-      assistantText(4),
-    ]);
-    const [intermediate, final] = assistants(out);
-    expect(isRegistered(intermediate)).toBe(false);
-    expect(isRegistered(final)).toBe(true);
-    expect(assistantSeqByRef.get(expectAssistant(final))).toBe(4);
-  });
-
-  it('registers the final assistant of every turn but not intermediate ones', () => {
-    const out = toAgentMessages([
-      user(1),
-      assistantText(2),
-      user(3),
-      assistantToolCall(4),
-      assistantText(5),
-    ]);
-    const [first, intermediate, last] = assistants(out);
-    expect(isRegistered(first)).toBe(true);
-    expect(assistantSeqByRef.get(expectAssistant(first))).toBe(2);
-    expect(isRegistered(intermediate)).toBe(false);
-    expect(isRegistered(last)).toBe(true);
-    expect(assistantSeqByRef.get(expectAssistant(last))).toBe(5);
-  });
-
-  it('registers the trailing tool-call-only assistant on an open turn', () => {
-    const out = toAgentMessages([user(1), assistantToolCall(2)]);
-    const [a] = assistants(out);
-    expect(isRegistered(a)).toBe(true);
-    expect(assistantSeqByRef.get(expectAssistant(a))).toBe(2);
-  });
-});
-
-describe('toAgentMessages — tool-result side-channel (#1718)', () => {
-  it('does NOT emit standalone ToolResultMessage entries into the display list', () => {
-    const out = toAgentMessages([
-      user(1),
-      assistantToolCall(2),
-      toolResult(3, 2),
-      assistantText(4),
-    ]);
-    expect(out.some((m) => m.role === 'toolResult')).toBe(false);
-  });
-
-  it('populates toolResultByCallId so the assistant renderer can pair results to calls', () => {
-    toAgentMessages([user(1), assistantToolCall(2), toolResult(3, 2), assistantText(4)]);
-    const tr = toolResultByCallId.get('tc-2');
-    expect(tr).toBeDefined();
-    expect(tr?.toolName).toBe('do_thing');
-    expect(tr?.isError).toBe(false);
-  });
-
-  it('marks kernel-style failure JSON results as errors', () => {
-    const errorResult: ChatMessageData = {
-      seq: 3,
-      role: 'tool_result',
-      content: '{"error": "boom"}',
-      tool_call_id: 'tc-2',
-      tool_name: 'do_thing',
-      created_at: ISO,
-    };
-    toAgentMessages([user(1), assistantToolCall(2), errorResult]);
-    expect(toolResultByCallId.get('tc-2')?.isError).toBe(true);
-  });
-
-  it('clears stale side-channel entries across conversions', () => {
-    toAgentMessages([user(1), assistantToolCall(2), toolResult(3, 2)]);
-    expect(toolResultByCallId.has('tc-2')).toBe(true);
-    toAgentMessages([user(1), assistantText(2)]);
-    expect(toolResultByCallId.has('tc-2')).toBe(false);
-  });
-
-  it('messagesForArtifactReconstruction re-weaves tool results after their paired assistant', () => {
-    const out = toAgentMessages([
-      user(1),
-      assistantToolCall(2),
-      toolResult(3, 2),
-      assistantText(4),
-    ]);
-    const woven = messagesForArtifactReconstruction(out);
-    const roles = woven.map((m) => m.role);
-    // Expect: user, assistant(toolCall), toolResult, assistant(text)
-    expect(roles).toEqual(['user', 'assistant', 'toolResult', 'assistant']);
-  });
-});
-
-describe('isFirstAssistantOfTurn (#1727)', () => {
-  it('flags the sole assistant after a user as first-of-turn', () => {
-    const out = toAgentMessages([user(1), assistantText(2)]);
-    const [a] = assistants(out);
-    expect(isFirstAssistantOfTurn(expectAssistant(a), out, hostsOf(out))).toBe(true);
-  });
-
-  it('flags the turn host of a multi-iteration turn, not the tool-call-only intermediate', () => {
-    // With per-turn chip cards (#1764) the tool-call-only intermediate is
-    // invisible — the avatar belongs to the host (last assistant) instead.
-    const out = toAgentMessages([user(1), assistantToolCall(2), assistantText(3)]);
-    const [first, second] = assistants(out);
-    expect(isFirstAssistantOfTurn(expectAssistant(first), out, hostsOf(out))).toBe(false);
-    expect(isFirstAssistantOfTurn(expectAssistant(second), out, hostsOf(out))).toBe(true);
-  });
-
-  it('treats intervening toolResult frames as transparent to turn position', () => {
-    const out = messagesForArtifactReconstruction(
-      toAgentMessages([user(1), assistantToolCall(2), toolResult(3, 2), assistantText(4)]),
-    );
-    const [first, second] = assistants(out);
-    expect(isFirstAssistantOfTurn(expectAssistant(first), out, hostsOf(out))).toBe(false);
-    expect(isFirstAssistantOfTurn(expectAssistant(second), out, hostsOf(out))).toBe(true);
-  });
-
-  it('treats an empty-thinking assistant frame as transparent so the next visible assistant owns the avatar', () => {
-    const emptyThinking: AssistantMessage = {
-      role: 'assistant',
-      content: [{ type: 'thinking', thinking: '' }],
-      usage: {
-        input: 0,
-        output: 0,
-        cacheRead: 0,
-        cacheWrite: 0,
-        totalTokens: 0,
-        cost: { input: 0, output: 0, cacheRead: 0, cacheWrite: 0, total: 0 },
-      },
-      stopReason: 'stop',
-      api: 'anthropic',
-      provider: 'anthropic',
-      model: 'claude-3-5-sonnet',
-      timestamp: Date.parse(ISO),
-    };
-    const visible = expectAssistant(assistants(toAgentMessages([user(1), assistantText(2)]))[0]);
-    const list = [...toAgentMessages([user(1)]), emptyThinking, visible];
-    expect(isFirstAssistantOfTurn(emptyThinking, list, hostsOf(list))).toBe(false);
-    expect(isFirstAssistantOfTurn(visible, list, hostsOf(list))).toBe(true);
-  });
-
-  it('restores first-of-turn after a new user message', () => {
-    const out = toAgentMessages([
-      user(1),
-      assistantToolCall(2),
-      assistantText(3),
-      user(4),
-      assistantText(5),
-    ]);
-    // Both turn hosts (text3, text5) are the visible first-of-turn frames;
-    // the tool-call-only intermediate is invisible.
-    const [first, second, third] = assistants(out);
-    expect(isFirstAssistantOfTurn(expectAssistant(first), out, hostsOf(out))).toBe(false);
-    expect(isFirstAssistantOfTurn(expectAssistant(second), out, hostsOf(out))).toBe(true);
-    expect(isFirstAssistantOfTurn(expectAssistant(third), out, hostsOf(out))).toBe(true);
-  });
-});
-
-describe('aggregateTurnToolCalls (#1764)', () => {
-  it('attaches every tool call from a single-iteration turn to that assistant', () => {
-    const out = toAgentMessages([
-      user(1),
-      assistantMultiToolCall(2, ['a', 'b', 'c']),
-      toolResult(3, 2, 'a'),
-      toolResult(4, 2, 'b'),
-      toolResult(5, 2, 'c'),
-    ]);
-    const [a] = assistants(out);
-    const agg = aggregateTurnToolCalls(out, toolResultByCallId);
-    const entries = agg.get(expectAssistant(a));
-    expect(entries).toBeDefined();
-    expect(entries?.map((e) => e.call.name)).toEqual(['a', 'b', 'c']);
-  });
-
-  it('collects tool calls across iterations onto the turn final assistant', () => {
-    const msgs = [
-      user(1),
-      assistantMultiToolCall(2, ['grep', 'read']),
-      toolResult(3, 2, 'grep'),
-      toolResult(4, 2, 'read'),
-      assistantToolCall(5, 'bash'),
-      toolResult(6, 5, 'bash'),
-      assistantText(7, 'done'),
-    ];
-    // The first tool_result seq must point to its call's seq; adjust.
-    msgs[2] = { ...(msgs[2] as ChatMessageData), tool_call_id: 'tc-2-0', tool_name: 'grep' };
-    msgs[3] = { ...(msgs[3] as ChatMessageData), tool_call_id: 'tc-2-1', tool_name: 'read' };
-    const out = toAgentMessages(msgs);
-    const as = assistants(out);
-    const agg = aggregateTurnToolCalls(out, toolResultByCallId);
-    // Final assistant hosts all three; others host none.
-    const last = expectAssistant(as[as.length - 1]);
-    const hosted = agg.get(last);
-    expect(hosted?.map((e) => e.call.name)).toEqual(['grep', 'read', 'bash']);
-    for (let i = 0; i < as.length - 1; i++) {
-      const intermediate = expectAssistant(as[i]);
-      expect(agg.get(intermediate)).toBeUndefined();
-    }
-  });
-
-  it('pairs tool results from the side-channel with their calls', () => {
-    const out = toAgentMessages([
-      user(1),
-      assistantToolCall(2),
-      toolResult(3, 2),
-      assistantText(4),
-    ]);
-    const agg = aggregateTurnToolCalls(out, toolResultByCallId);
-    const last = assistants(out)[1];
-    const entries = agg.get(expectAssistant(last));
-    expect(entries?.length).toBe(1);
-    expect(entries?.[0]?.result).toBeDefined();
-    expect(entries?.[0]?.result?.toolCallId).toBe('tc-2');
-  });
-
-  it('scopes each turn’s card to its own tool calls across a multi-turn sequence', () => {
-    const msgs = [
-      user(1),
-      assistantToolCall(2, 'alpha'),
-      toolResult(3, 2, 'alpha'),
-      assistantText(4, 'first reply'),
-      user(5),
-      assistantToolCall(6, 'beta'),
-      toolResult(7, 6, 'beta'),
-      assistantText(8, 'second reply'),
-    ];
-    const out = toAgentMessages(msgs);
-    const agg = aggregateTurnToolCalls(out, toolResultByCallId);
-    const as = assistants(out);
-    // Turn 1 host is the second assistant (text) — tc-2 only.
-    // Turn 2 host is the fourth assistant (text) — tc-6 only.
-    const host1 = expectAssistant(as[1]);
-    const host2 = expectAssistant(as[3]);
-    expect(agg.get(host1)?.map((e) => e.call.name)).toEqual(['alpha']);
-    expect(agg.get(host2)?.map((e) => e.call.name)).toEqual(['beta']);
-  });
-});
